Refactor src -> _src
