+++ python/format/_src/errors_test.py
@@ -2,7 +2,7 @@
 
 import textwrap
 
-from format.src import errors
+from format._src import errors
 
 
 def test_issues():
