+++ python/format/_src/tests/graphs/metadata_bad_type.json
