+++ python/format/_src/tests/graphs/metadata_missing_property_name.json
